+++ rust/pyxel-platform/build.rs
@@ -63,6 +63,25 @@ impl SDL2BindingsBuilder {
             }
             panic!("Failed to extract SDL2 source code");
         }
+
+        // Patch SDL2 for macOS Sonoma
+        if self.target_os == "darwin" {
+            let patch_target_path = format!(
+                "{}/SDL2-{}/src/video/cocoa/SDL_cocoaevents.m",
+                self.out_dir, SDL2_VERSION
+            );
+            let patch_code =
+            "- (BOOL)applicationSupportsSecureRestorableState:(NSApplication *)app { return YES; }";
+            let status = Command::new("sh")
+                .arg("-c")
+                .arg(&format!(
+                    "sed -i '' '310i\\\n{}' {}",
+                    patch_code, patch_target_path
+                ))
+                .status()
+                .expect("Failed to execute sed command");
+            assert!(status.success(), "Failed to patch SDL2");
+        }
     }
 
     fn build_sdl2(&self) {